requirements.txt

gunicorn==0.16.1
django-image-cropping
easy_thumbnails
django-admin-sortable
django-suit



